Move messaging protocol buffers definitiosn into `draftspecs` package.
