+++ internal/draftspecs/doc.go
@@ -1,0 +1,3 @@
+// Package draftspecs contains draft specifications for protocol-buffers formats
+// that will allow for interoperation across Dogma engine implementations.
+package draftspecs
